… views/stockoverview.blade.php