Hi,

I tried to improve the Performance of the "Stock overview" page, and found some improvements:
1) I replaced some of the `FindObjectInArrayByPropertyValue`-calls by getting this values directly in the database query.
This also made the template-code more readable.

2) I optimized the "stock_current"-database-view:
I created 2 indices, and replaced a subquery with a (left-)join.
I'm not sure why, but this combination improved the speed significantly.
stock_current is used often:
I got a speedup from 1.26s to 0.13s on /stock/products #848
The view stock_missing_products, which uses stock_current, is now 25 times faster (from ~1250 ms to ~50ms) on my sample data.


Might be related to #898 (Note that JavaScript ("Datatables") also has some loading time. I think pagination could fix that.)