@berrnd 2020-07-24
Great, thanks a lot for that too! :)

Makes sense to have the columns needed for the stock overview table populated by a SQL query - just one thing:
I tried to avoid to have "bigger" SQL queries in-code.
So maybe a view instead of the query in `GetCurrentStockOverview()` would be great, if possible somehow with the dynamic JOIN part based on feature flags...

> Note that JavaScript ("Datatables") also has some loading time. I think pagination could fix that.

The only real enhancement would be to have pagination + load the data dynamically and not render the whole HTML table server side...
The biggest table the stock journal, I have about 10K lines there on my private instances which loads currently in about 10 seconds - great that this page is only needed rarely on daily grocy use cases... :D