@fipwmaqzufheoxq92ebc 2020-07-25
> Makes sense to have the columns needed for the stock overview table populated by a SQL query - just one thing:
> I tried to avoid to have "bigger" SQL queries in-code.
> So maybe a view instead of the query in `GetCurrentStockOverview()` would be great, if possible somehow with the dynamic JOIN part based on feature flags...

I changed that: I create now two views and query them.