@berrnd 2020-08-18
Thanks again - the page speed improvement on the stock overview page is noticeable. :)

Maybe we should extend this "helper-view-approach" to at least all other overview pages...