@berrnd 2020-11-10
I have no explanation, but any of the indexes create in [that migration](https://github.com/grocy/grocy/blob/62e8d88adb321d916a1a61eccb4efaf081f7f354/migrations/0106.sql) made that the `stock_current` view failed (the `is_aggregated_amount` did not work anymore, returned always `0`).

I searched for over an hour, but haven't found out what specific thing the problem is, it seems to be the combination of those indexes - strange...